pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yt-seo-tool"
version = "0.1.0"
description = "YouTube SEO Refresh Tool"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12",
  "streamlit>=1.28",
  "tomli>=2; python_version<'3.11'",
  "yt-dlp>=2024.10.7",
  "requests>=2.31",
  "google-auth>=2.23",
  "google-auth-oauthlib>=1.1",
  "google-api-python-client>=2.108"
]

[project.scripts]
ytseo = "cli.main:main"

[tool.setuptools.packages.find]
include = ["ytseo*", "cli*"]
